feat(zr_vm): AOT slot cache, LLVM value layout fix, typed-opcode wave-2, GC, LSP, benchmarks
- AOT runtime: cache generated frame slot counts; remove hot-path rescan of
  ZrCore_Function_GetGeneratedFrameSlotCount; tighten BeginInstruction with a
  non-observing fast path while preserving observable semantics
- AOT LLVM: fix SZrTypeValue / SZrTypeValueOnStack lowering to match C runtime
  alignment (opaque stack blobs + byte-offset access) and repair related
  arithmetic/member lowering paths
- Compiler/parser: extend typed opcode, closure, and native-binding paths across
  interpreter and AOT C/LLVM backends; refresh compiler integration and writer output
- VM core & GC: update execution dispatch, functions/closures, I/O, exceptions,
  and GC mark/cycle/object handling; adjust zr_vm_common instruction/object/runtime config
- zr.system / FFI: align GC registry and filesystem helpers with core GC semantics
- Language server & VS Code extension: semantic/type prototype support, diagnostics,
  and native/WASM asset sync scripts
- Tests & tooling: new/updated benchmark cases (GC fragmentation, typed equality,
  known call, etc.), golden AST/intermediate/runtime for classes, matrix fixtures,
  CMake test host env helpers, perf/benchmark scripts, and acceptance notes
- Docs: system submodules and CTest performance reporting updates

+++ tests/ffi/ffi_fixture.c
@@ -1,6 +1,7 @@
 #include <stddef.h>
 #include <stdarg.h>
 #include <stdint.h>
+#include <stdio.h>
 #include <string.h>
 
 #if defined(_WIN32)
@@ -195,8 +196,8 @@ ZR_FFI_FIXTURE_EXPORT int32_t zr_ffi_flip_mode(int32_t modeValue) {
 
 ZR_FFI_FIXTURE_EXPORT int32_t zr_ffi_tell_fd(int32_t fd) {
 #if defined(_WIN32)
-    return (int32_t)_lseeki64(fd, 0, SEEK_CUR);
+    return (int32_t)_lseeki64((int)fd, 0, SEEK_CUR);
 #else
-    return (int32_t)lseek(fd, 0, SEEK_CUR);
+    return (int32_t)lseek((int)fd, 0, SEEK_CUR);
 #endif
 }
